+++ modules/planning/optimizer/qp_spline_path/qp_frenet_frame.cc
@@ -38,7 +38,7 @@ bool QpFrenetFrame::Init(const ReferenceLine& reference_line,
                          const common::FrenetFramePoint& init_frenet_point,
                          const double start_s, const double end_s,
                          const double time_resolution,
-                         const std::size_t num_points) {
+                         const std::uint32_t num_points) {
   _vehicle_param = common::config::VehicleConfigHelper::instance()
                        ->GetConfig()
                        .vehicle_param();
@@ -70,7 +70,7 @@ bool QpFrenetFrame::Init(const ReferenceLine& reference_line,
 
   const double inf = std::numeric_limits<double>::infinity();
 
-  for (std::size_t i = 0; i <= num_points; ++i) {
+  for (std::uint32_t i = 0; i <= num_points; ++i) {
     double s = _start_s + i * resolution;
     _evaluated_knots.push_back(std::move(s));
     _hdmap_bound.push_back(std::make_pair(-inf, inf));
@@ -182,7 +182,8 @@ bool QpFrenetFrame::mapping_dynamic_obstacle_with_decision(
     return false;
   }
 
-  for (std::size_t i = 0; i < obstacle.prediction_trajectories().size(); ++i) {
+  for (std::uint32_t i = 0; i < obstacle.prediction_trajectories().size();
+       ++i) {
     if (!decision[i].has_nudge()) {
       continue;
     }
@@ -206,7 +207,7 @@ bool QpFrenetFrame::mapping_dynamic_obstacle_with_decision(
       obs_box.GetAllCorners(&corners);
       std::vector<common::SLPoint> sl_corners;
 
-      for (std::size_t i = 0; i < corners.size(); ++i) {
+      for (std::uint32_t i = 0; i < corners.size(); ++i) {
         common::SLPoint cur_point;
         Eigen::Vector2d xy_point(corners[i].x(), corners[i].y());
         if (!_reference_line->get_point_in_Frenet_frame(xy_point, &cur_point)) {
@@ -220,7 +221,7 @@ bool QpFrenetFrame::mapping_dynamic_obstacle_with_decision(
         sl_corners.push_back(std::move(cur_point));
       }
 
-      for (std::size_t i = 0; i < sl_corners.size(); ++i) {
+      for (std::uint32_t i = 0; i < sl_corners.size(); ++i) {
         common::SLPoint sl_first = sl_corners[i % sl_corners.size()];
         common::SLPoint sl_second = sl_corners[(i + 1) % sl_corners.size()];
         if (sl_first.s() < sl_second.s()) {
@@ -242,10 +243,10 @@ bool QpFrenetFrame::mapping_dynamic_obstacle_with_decision(
             update_start_s < _evaluated_knots.front()) {
           continue;
         }
-        std::pair<std::size_t, std::size_t> update_index_range =
+        std::pair<std::uint32_t, std::uint32_t> update_index_range =
             find_interval(update_start_s, update_end_s);
 
-        for (std::size_t j = update_index_range.first;
+        for (std::uint32_t j = update_index_range.first;
              j < update_index_range.second; ++j) {
           _dynamic_obstacle_bound[j].first =
               std::max(bound.first, _dynamic_obstacle_bound[j].first);
@@ -300,7 +301,7 @@ bool QpFrenetFrame::mapping_polygon(
     std::vector<std::pair<double, double>>* const bound_map) {
   std::vector<common::SLPoint> sl_corners;
 
-  for (std::size_t i = 0; i < corners.size(); ++i) {
+  for (std::uint32_t i = 0; i < corners.size(); ++i) {
     common::SLPoint cur_point;
     Eigen::Vector2d xy_point(corners[i].x(), corners[i].y());
     if (!_reference_line->get_point_in_Frenet_frame(xy_point, &cur_point)) {
@@ -313,7 +314,7 @@ bool QpFrenetFrame::mapping_polygon(
     sl_corners.push_back(std::move(cur_point));
   }
 
-  for (std::size_t i = 0; i < sl_corners.size(); ++i) {
+  for (std::uint32_t i = 0; i < sl_corners.size(); ++i) {
     if (!map_line(sl_corners[i % sl_corners.size()],
                   sl_corners[(i + 1) % sl_corners.size()], nudge_side,
                   bound_map)) {
@@ -331,7 +332,7 @@ bool QpFrenetFrame::map_line(
     std::vector<std::pair<double, double>>* const constraint) {
   common::SLPoint near_point = (start.s() < end.s() ? start : end);
   common::SLPoint further_point = (start.s() < end.s() ? end : start);
-  std::pair<std::size_t, std::size_t> impact_index =
+  std::pair<std::uint32_t, std::uint32_t> impact_index =
       find_interval(near_point.s(), further_point.s());
 
   if ((near_point.s() < _start_s - _vehicle_param.back_edge_to_center() &&
@@ -345,7 +346,7 @@ bool QpFrenetFrame::map_line(
       std::max(std::abs(near_point.s() - further_point.s()), 1e-8);
   double weight = 0.0;
 
-  for (std::size_t i = impact_index.first; i <= impact_index.second; ++i) {
+  for (std::uint32_t i = impact_index.first; i <= impact_index.second; ++i) {
     weight = std::abs((_evaluated_knots[i] - near_point.s())) / distance;
     weight = std::max(weight, 0.0);
     weight = std::min(weight, 1.0);
@@ -414,14 +415,14 @@ std::pair<double, double> QpFrenetFrame::map_lateral_constraint(
   return result;
 }
 
-std::pair<std::size_t, std::size_t> QpFrenetFrame::find_interval(
+std::pair<std::uint32_t, std::uint32_t> QpFrenetFrame::find_interval(
     const double start, const double end) const {
   double new_start = std::max(start - _vehicle_param.front_edge_to_center(),
                               _evaluated_knots.front());
   double new_end = std::min(end + _vehicle_param.back_edge_to_center(),
                             _evaluated_knots.back());
-  std::size_t start_index = find_index(new_start);
-  std::size_t end_index = find_index(new_end);
+  std::uint32_t start_index = find_index(new_start);
+  std::uint32_t end_index = find_index(new_end);
 
   if (end > _evaluated_knots[end_index] &&
       end_index + 1 != _evaluated_knots.size()) {
@@ -432,7 +433,7 @@ std::pair<std::size_t, std::size_t> QpFrenetFrame::find_interval(
 }
 
 bool QpFrenetFrame::calculate_hd_map_bound() {
-  for (std::size_t i = 0; i < _hdmap_bound.size(); ++i) {
+  for (std::uint32_t i = 0; i < _hdmap_bound.size(); ++i) {
     _hdmap_bound[i].first = -4.0;
     _hdmap_bound[i].second = 4.0;
   }
@@ -443,7 +444,7 @@ bool QpFrenetFrame::calculate_static_obstacle_bound() {
   const std::vector<const Obstacle*> static_obs_list =
       _decision_data->StaticObstacles();
 
-  for (std::size_t i = 0; i < static_obs_list.size(); ++i) {
+  for (std::uint32_t i = 0; i < static_obs_list.size(); ++i) {
     const Obstacle* cur_obstacle = static_obs_list[i];
     if (mapping_static_obstacle_with_decision(*cur_obstacle)) {
       AERROR << "mapping obstacle with id [" << cur_obstacle->Id()
@@ -458,7 +459,7 @@ bool QpFrenetFrame::calculate_dynamic_obstacle_bound() {
   const std::vector<const Obstacle*> dynamic_obs_list =
       _decision_data->DynamicObstacles();
 
-  for (std::size_t i = 0; i < dynamic_obs_list.size(); ++i) {
+  for (std::uint32_t i = 0; i < dynamic_obs_list.size(); ++i) {
     const Obstacle* cur_obstacle = dynamic_obs_list[i];
     if (mapping_dynamic_obstacle_with_decision(*cur_obstacle)) {
       AERROR << "mapping obstacle with id [" << cur_obstacle->Id()
@@ -481,7 +482,7 @@ bool QpFrenetFrame::get_bound(
   }
 
   // linear bound interpolation
-  std::size_t lower_index = find_index(s);
+  std::uint32_t lower_index = find_index(s);
   const double s_low = _evaluated_knots[lower_index];
   const double s_high = _evaluated_knots[lower_index + 1];
   double weight = (Double::compare(s_high, s_low, 1e-8) <= 0
@@ -514,12 +515,12 @@ bool QpFrenetFrame::get_bound(
   return true;
 }
 
-std::size_t QpFrenetFrame::find_index(const double s) const {
+std::uint32_t QpFrenetFrame::find_index(const double s) const {
   auto upper_bound =
       std::lower_bound(_evaluated_knots.begin() + 1, _evaluated_knots.end(), s);
-  return std::min(
-             _evaluated_knots.size() - 1,
-             static_cast<std::size_t>(upper_bound - _evaluated_knots.begin())) -
+  return std::min(static_cast<std::uint32_t>(_evaluated_knots.size() - 1),
+                  static_cast<std::uint32_t>(upper_bound -
+                                             _evaluated_knots.begin())) -
          1;
 }
